% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnFamilyBasedPGMs.R
\name{calculateBlockDiagonal2PhiMatrix}
\alias{calculateBlockDiagonal2PhiMatrix}
\title{Kinship matrix of a known pedigree structure}
\usage{
calculateBlockDiagonal2PhiMatrix(ped, squaredRoot = FALSE,
  sampled = NULL)
}
\arguments{
\item{ped}{A data.frame with with the pedigrees of the families, with columuns
\code{famid}, \code{id}, \code{dadid}, \code{momid}, and \code{sex} for all sampled
and non-sampled subjects.}

\item{squaredRoot}{a logical value indicating if the square root of the
kinship matrix  \eqn{2 \bm\Phi} (i.e., the Z matrix) must be computed.}

\item{sampled}{A logical vector in which element i indicates whether individual i was sampled or not.}
}
\value{
The kinship matrix \eqn{2 \bm\Phi} or its squared root, if squaredRoot is TRUE.
}
\description{
Computes the block diagonal kinship matrix with the degree of relatedness
between individuals all individuals of \eqn{F} families, multiplied by 2.
It is usually denoted as \eqn{2 \bm\Phi}, in which \eqn{\bm \Phi = diag\{\Phi^{(f)}, f=1,\ldots,F\}} and
each entry \eqn{\Phi^{(f)}_{ij}} is the probability that two alleles sampled at random from individuals
\eqn{i} and \eqn{j} of family \eqn{f} are identical by descent.
Thus, when the pedigree structure is known, \eqn{\Phi^{(f)}_{ii} = 1/2},
\eqn{\Phi^{(f)}_{ij} = 1/4} if \eqn{i} and \eqn{j} are siblings or if one of them is a parent of the other,
\eqn{\Phi^{(f)}_{ij} = 1/8} if one of them is a grand-parent of the other, and so on
\insertCite{lange2003mathematical}{FamilyBasedPGMs}.
}
