% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnFamilyBasedPGMs.R
\name{familyBasedCITest}
\alias{familyBasedCITest}
\title{Conditional Independence Test for Family Data}
\usage{
familyBasedCITest(x, y, S, suffStat)
}
\arguments{
\item{x}{An integer value indicating the position of variable X.}

\item{y}{An integer value  indicating the position of variable Y.}

\item{S}{An integer vector with the positions of zero or more conditioning variables in S.}

\item{suffStat}{A list with the elements "phen.df", "covs.df", "pedigrees",
"minK", "maxFC", "orthogonal", "alpha", "dirToSave", "fileID", "savePlots", and
"useGPU", as described in function \code{learnFamilyBasedDAGs}, and also with the
elements "Phi2" and "Z", both obtained by the function
\code{calculateBlockDiagonal2PhiMatrix} with the argument "squaredRoot" set,
respectively, as FALSE and TRUE.}
}
\value{
The p-value of the test.
}
\description{
Computes the p-value of the test of the null hypothesis of zero genetic, environmental,
or total partial correlation between X and Y given S, if dagg, dage, or dagt is set as TRUE, respectively.

The unconfounded estimation and significance test for the genetic and environmental partial correlation
coefficients are described in detail in \insertCite{ribeiro2019family;textual}{FamilyBasedPGMs}.
They are based on univariate polygenic linear mixed models \insertCite{almasy1998multipoint}{FamilyBasedPGMs},
with two components of variance: the polygenic or family-specific random effect,
which models the phenotypic variability across the families,
and the environmental or subject-specific error, which models phenotypic variability after removing the
familial aggregation effect.

This function can be used as a d-separation oracle in causal structure learning algorithms such as the
IC/PC algorithm when the variables are normal distributed and the observations are clusterized in families.
}
