% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnFamilyBasedPGMs.R
\name{learnFamilyBasedDAGs}
\alias{learnFamilyBasedDAGs}
\title{Learning of Gaussian Directed Acyclic PGMs from Family Data}
\usage{
learnFamilyBasedDAGs(phen.df, covs.df, pedigrees, sampled, fileID,
  dirToSave, alpha = 0.05, max_cores = NULL, minK = 10,
  maxFC = 0.01, orthogonal = TRUE, maj.rule = TRUE, useGPU = FALSE,
  debug = TRUE, savePlots = FALSE)
}
\arguments{
\item{phen.df}{A data.frame with phenotype variables of only sampled subjects.
Column names must be properly set with the names of the phenotypes.}

\item{covs.df}{A data.frame with covariates of only sampled subjects.
Column names must be properly set with the names of the covariates.}

\item{pedigrees}{A data.frame with columuns \code{famid}, \code{id}, \code{dadid},
\code{momid}, and \code{sex} columns for all sampled and non-sampled subjects.}

\item{sampled}{A logical vector in which element i indicates whether individual i was sampled or not.}

\item{fileID}{A character string to be used as prefix in the filenames of RData
objects with the partial correlation results. Note that covariates are not
identified in these files.}

\item{dirToSave}{Path to the folder you want to save the output objects.}

\item{alpha}{The significance level to be used in the partial correlation tests.}

\item{max_cores}{An integer indicating the maximum number of CPU cores
to be used for parallel execution.}

\item{minK}{A scalar indicating the minimum dimension allowed
in the dimensionality reduction for confounding correction.}

\item{maxFC}{A scalar between 0 and 1, indicating the maximum fraction of
confounding allowed.}

\item{orthogonal}{A logical value indicating whether the transformation matrix used in the
confounding correction is orthogonal or not.}

\item{maj.rule}{A logical value to be used in the \code{pc} function indicating wheter the
majority rule must be applied or not.}

\item{useGPU}{A logical value indicating whether GPU cores can be used for parallel execution.}

\item{debug}{A logical value indicating whether some debug messages can be shown.}

\item{savePlots}{A logical value indicating whether plots for the confounding correction must be generated.}
}
\value{
Returns a list with the partial correlation matrices (\code{pcor}), the
adjacency matrices (\code{adjM}), and with the igraph objects representing the
undirected PGM (\code{udg}).
}
\description{
The CPDAG representing the Markov equivalence classes to which the Gaussian directed acyclic
PGM belong and its decomposition into genetic and environmental components are learned
from observational family data by applying the IC/PC algorithm
\insertCite{pearl2000causality,spirtes2000causation}{FamilyBasedPGMs}
with the zero partial correlation tests derived in the work by
\insertCite{ribeiro2019family;textual}{FamilyBasedPGMs} as d-separation oracles.

These tests are based on univariate polygenic linear mixed models
\insertCite{almasy1998multipoint}{FamilyBasedPGMs}, with two components of variance: the polygenic or
family-specific random effect, which models the phenotypic variability across the families,
and the environmental or subject-specific error, which models phenotypic variability after removing the
familial aggregation effect.
}
\examples{
\donttest{
data(scen3) # available simulated datasets are scen1, scen2, scen3, and scen4

scenario = 3 # data was simulated according to scenario 3

fam.nf <- scen3$fam.nf
pedigrees <- scen3$pedigrees
phen.df <- scen3$phen.df[[1]] # accessing the first replicate
covs.df <- NULL # no covariates were used in the simulation process.

N <- sum(fam.nf) # total number of individuals
sampled <- rep(1, N) # in simulated data, all individuals were sampled.

fileID <- paste0("scen", scenario)
dirToSave <- paste0("./objects-PC-", fileID, "/")
dir.create(dirToSave, showWarnings=FALSE)

alpha = 0.05

dags <- learnFamilyBasedDAGs(phen.df, covs.df, pedigrees, sampled,
                             fileID, dirToSave, alpha, max_cores=NULL,
                             minK=10, maxFC = 0.05, orthogonal=TRUE, maj.rule=TRUE,
                             useGPU=FALSE, debug=TRUE, savePlots=FALSE)

# the adjacency matrix of the learned directed acyclic genetic PGM
as(dags$g, "amat")

# plotting the the learned directed acyclic genetic PGM as an `igraph` object:
plot.igraph(graph.adjacency(adjM_g), vertex.size=30, vertex.color="lightblue")

# the adjacency matrix of the learned directed acyclic environmental PGM
as(dags$e, "amat")

# plotting the the learned directed acyclic environmental PGM as an `igraph` object:
plot.igraph(graph.adjacency(adjM_e), vertex.size=30, vertex.color="lightblue")
}

}
\references{
{
   \insertAllCited
}
}
