% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePedigree.R
\name{plotFamilyPedigree}
\alias{plotFamilyPedigree}
\title{Plotting of Pedigree Chart}
\usage{
plotFamilyPedigree(pedigrees, famid)
}
\arguments{
\item{pedigrees}{A data.frame object with at least the columns: "famid", "id", "momid", "dadid", and "sex"}

\item{famid}{A integer value identifying the family for which the pedigree chart should be plotted.}
}
\description{
Creates a chart of the pedigree of family \code{famid}.
}
\examples{
\donttest{
  data(scen1)
  pedigrees <- scen1$pedigrees

  # plot the pedigree chart of the fifth simulated family
  plotFamilyPedigree(pedigrees, 5)
}

}
