% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePedigree.R
\name{simulatePedigrees}
\alias{simulatePedigrees}
\title{Pedigree Simulation}
\usage{
simulatePedigrees(fam.nf, nInitFounderCouples, probToMarry, probChildren)
}
\arguments{
\item{fam.nf}{A integer vector where the entry i indicates the number of individuals at family i.}

\item{nInitFounderCouples}{A integer value representing the number of founders in the first generation.}

\item{probToMarry}{A scalar between 0 and 1 indicating the probability of a individual getting married.
Marriages only occur between individuals of the same generation.}

\item{probChildren}{A vector with values between 0 and 1 in which entry i indicates the probability
of a couple has i children.}
}
\description{
Pedigree Simulation
}
\examples{
fam.nf <- rep(10, 20)  # for 20 families with 10 individuals
nInitFounderCouples <- 1
probChildren <- c(0, 0, 0.3, 0.3, 0.3, 0.1)
probToMarry = 0.85
pedigrees <- simulatePedigrees(fam.nf, nInitFounderCouples, probToMarry, probChildren)

}
